+++ backend/footprint/src/main/java/foot/footprint/domain/article/application/findArticleDetails/FindPublicArticleDetailsService.java
@@ -1,7 +1,7 @@
 package foot.footprint.domain.article.application.findArticleDetails;
 
 import foot.footprint.domain.article.dao.FindArticleRepository;
-import foot.footprint.domain.article.dto.ArticlePageResponse;
+import foot.footprint.domain.article.dto.articleDetails.ArticlePageResponse;
 import foot.footprint.domain.articleLike.dao.ArticleLikeRepository;
 import foot.footprint.domain.comment.dao.FindCommentRepository;
 import foot.footprint.domain.commentLike.dao.CommentLikeRepository;
