+++ backend/footprint/src/main/java/foot/footprint/domain/article/dto/articleDetails/ArticleDetailsDto.java
@@ -1,4 +1,4 @@
-package foot.footprint.domain.article.dto;
+package foot.footprint.domain.article.dto.articleDetails;
 
 import java.util.Date;
 import lombok.AllArgsConstructor;
