+++ backend/footprint/src/main/java/foot/footprint/domain/article/dto/articleDetails/ArticlePageResponse.java
@@ -1,4 +1,4 @@
-package foot.footprint.domain.article.dto;
+package foot.footprint.domain.article.dto.articleDetails;
 
 import foot.footprint.domain.comment.dto.CommentResponse;
 import foot.footprint.domain.comment.dto.CommentsDto;
